+++ aws_emr_launch/__init__.py
@@ -11,3 +11,6 @@
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
 
+import pkg_resources
+
+__version__ = pkg_resources.get_distribution("aws-emr-launch").version